fix(codegen,linker): assembler/encoder/linker hardening pass
AArch64 assembler:
- chooseGprScratch() cycles {X9,X10,X11} avoiding base and optional src
  to prevent register corruption in large-offset load/store expansions.
  All four FP/GPR emit helpers now use it instead of hardcoded kScratchGPR.
- A64BinaryEncoder: checkedPairImm7() validates LDP/STP pair offsets for
  8-byte alignment and signed imm7 range before binary encoding.
  encodeLargeOffsetLdSt() picks scratch candidates avoiding base+dest.

x86-64 encoder/unwind:
- EmittedWin64UnwindOp struct tracks win64 unwind ops by descriptor
  rather than fragile prologue-instruction index arithmetic.
  instrMatchesWin64UnwindOp() verifies the emitted MIR instruction
  matches its declared unwind role before recording it.

Linker — relocation resolution:
- RelocApplier: local symbols (sectionIndex > 0 || absolute) resolved
  before falling through to global name lookup; eliminates stale-global
  shadowing for locally-defined symbols.
- BranchTrampoline: resolveRelocSymbol() helper unifies local+global
  resolution in the same priority order as RelocApplier.
- ICF: uses classifyReloc() for address-taken detection; executable
  sections are now scanned for non-branch relocations (ADRP+ADD,
  RIP-relative LEA) so address-taken functions are not folded.

Linker — object readers:
- CoffReader/ElfReader/MachOReader: replace offset+sizeof(T)>size with
  overflow-safe offset>size||sizeof(T)>size-offset pattern.
  checkedRange() used consistently for addend extraction.
- All three readers: null symbol pre-allocated with assign(1,{}) at
  parse start so symbol table is always valid even on early error paths.
- MachOReader: LC_SYMTAB out-of-bounds is now a hard error; missing
  LC_SYMTAB with relocations is rejected instead of silently ignored.

Linker — section merger / dead strip:
- SectionMerger: sectionClassOrder() provides a stable numeric ordering
  for SectionClass enum values; stable_sort now orders chunks by class
  before alignment-descending sort within class.
- DeadStripPass: .init_array.N, .fini_array.N, .preinit_array* forms
  are always-live (priority-suffixed ELF init/fini arrays).

Object writers — unified reloc resolution + rodata relocs:
- CoffWriter/ElfWriter/MachOWriter: duplicate rodata symbol names now
  marked with sentinel UINT32_MAX so ambiguous cross-section refs fall
  back to the source-map lookup instead of silently picking the first.
  Shared resolveRelocSym lambda replaces three divergent ad-hoc paths.
- ElfWriter: adds .rela.rodata section; rodata-to-text and
  rodata-to-rodata relocations are now emitted (jump tables, vtables,
  const function-pointer arrays were previously dropped silently).
  Section indices bumped: kSecRelaRodata=4, symtab=5, strtab=6, etc.
- CoffWriter: .rdata now gets its own relocation table, overflow record,
  and section header flags (IMAGE_SCN_LNK_NRELOC_OVFL when needed).
- MachOWriter: __data addend relocs use ARM64_RELOC_ADDEND (10) instead
  of being dropped.
- StringDedup: dedup symbol names changed from __dedup_str_N to
  __viper_dedup_str_N with collision-free generation loop checking both
  global symbol table and per-object symbol tables.
- ElfExeWriter: dynInfoForSym return type fixed uint32_t→uint64_t.

Tests: new coverage for all hardening paths across binary encoders,
       linker passes (BranchTrampoline, DeadStrip, ICF, RelocApplier,
       SectionMerger, StringDedup), and all three object writers.

+++ src/codegen/common/linker/ElfExeWriter.cpp
@@ -235,7 +235,7 @@ uint32_t addString(std::vector<uint8_t> &strtab, const std::string &s) {
     return off;
 }
 
-uint32_t dynInfoForSym(uint32_t symIndex, uint32_t type) {
+uint64_t dynInfoForSym(uint32_t symIndex, uint32_t type) {
     return (static_cast<uint64_t>(symIndex) << 32) | type;
 }
